added better ml models for stock prediction

+++ MachineLearningModel/Model.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
 using System.IO;
+using System.Linq;
 using Microsoft.ML;
 using Microsoft.ML.Transforms.TimeSeries;
 
@@ -21,17 +22,19 @@ namespace MachineLearningModel
             MessageId = "type: System.Reflection.CustomAttributeNamedParameter[]; size: 245MB")]
         public static void CreateAndSaveModel(string stock)
         {
-            var dataPath = Path.Combine(Environment.CurrentDirectory, "StockData", "stocks", String.Concat(stock, ".csv"));
+            var dataPath = Path.Combine(Environment.CurrentDirectory, "StockData", "stocks",
+                String.Concat(stock, ".csv"));
             Console.WriteLine(dataPath);
+            var trainSize = File.ReadLines(dataPath).Count() - 1;
             var data = _mlContext.Data.LoadFromTextFile<StockData>(dataPath, hasHeader: true, separatorChar: ',');
             var pipeline = _mlContext.Forecasting.ForecastBySsa(
                 "HighPricePredicted",
                 nameof(StockData.HighPrice),
-                20,
+                Math.Min(trainSize / 3 + 2, 100),
+                Math.Min(trainSize / 2, 150),
+                trainSize - 1,
                 100,
-                5000,
-                100,
-                confidenceLevel: .98f);
+                confidenceLevel: 0.95f);
             var model = pipeline.Fit(data);
             _mlContext.Model.Save(model, data.Schema,
                 Path.Combine(Environment.CurrentDirectory, "StockData", "Models", String.Concat(stock, ".zip")));
@@ -59,11 +62,26 @@ namespace MachineLearningModel
             {
                 if (file.Contains(".zip"))
                 {
-                    File.Delete(file);
-                    continue;
+                    // File.Delete(file);
                 }
-                var stock = file.Replace(".csv", "");
-                Model.CreateAndSaveModel(stock);
+                else
+                {
+                    var filePath = file.Replace(".csv", ".zip");
+                    if (File.Exists(filePath))
+                    {
+                        continue;
+                    }
+
+                    try
+                    {
+                        var stock = file.Replace(".csv", "");
+                        Model.CreateAndSaveModel(stock);
+                    }
+                    catch (Exception e)
+                    {
+                        File.Delete(file);
+                    }
+                }
             }
         }
     }
